% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onedrive.R
\name{get_onedrive}
\alias{get_onedrive}
\title{get_onedrive}
\usage{
get_onedrive()
}
\value{
A string containing the file location of the OneDrive folder
}
\description{
get_onedrive
}
\examples{
get_onedrive()

}
